Bug 1844830 - patch 3 - gfxFontGroup::GetFontAt should take an unsigned index. r=gfx-reviewers,lsalzman
There's no reason for this to be signed, and the callers are using an unsigned index already.

(Won't make any difference, especially once the optimizer has had a look at it; this is just
a bit of cleanup noticed in passing.)

Differential Revision: https://phabricator.services.mozilla.com/D184313

+++ gfx/thebes/gfxTextRun.cpp
@@ -2011,9 +2011,9 @@ bool gfxFontGroup::HasFont(const gfxFontEntry* aFontEntry) {
   return false;
 }
 
-already_AddRefed<gfxFont> gfxFontGroup::GetFontAt(int32_t i, uint32_t aCh,
+already_AddRefed<gfxFont> gfxFontGroup::GetFontAt(uint32_t i, uint32_t aCh,
                                                   bool* aLoading) {
-  if (uint32_t(i) >= mFonts.Length()) {
+  if (i >= mFonts.Length()) {
     return nullptr;
   }
